feat: implementa ContatoController e ContatoRepository

+++ agenda/src/main/java/com/agenda/agenda/repository/ContatoRepository.java
@@ -2,5 +2,5 @@ package com.agenda.agenda.repository;
 import com.agenda.agenda.model.Contato;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContatoRepository extends JpaRepository<Contato, Integer> {
+public interface ContatoRepository extends JpaRepository<Contato, Long> {
 }
